Jekyll web created for the notes
I used Jekyll and the Odin theme to create the web, which uses a card
style template and sass for the styles.

The markdown files have been moved to the folder "_posts", the images to
the folder "assets" and the jupyter notebooks to the folder "_jupyter".

Every card of the index belong to a lesson of the Computer Vision
Nanodegree of Udacity that I'm taking.

The jupyter notebooks haven't been converted to markdown files yet and there are still some design issues.