Solving the problem with the baseurl
I found the problem in the _config.yml file. At the beginning the base
url was "computer-vision" and that was causing that the url was
"localhost:4000computer-vision". So I only had to change it for
"/computer-vison" and now is working.

I also had to change some "a" tags because the Jekyll theme add
automatically a slash at the end of the url: "/computer-vision/".

Besides, I converted the logo of the footer into an "a" tag to redirect
the user to the home.