Jupyter to md program & Pytorch section created
To test the program you have to execute in the terminal "python3
jupyter2md.py" and it will pop up a window to select the file. After
that, it'll move the created markdown and images to their respective
folders.

I also made some changes in the styles and continued with the fifth
section.